+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add Samsung device detection.
 - Delay parsing `Accept-Language` until `Browser::Base#accept_language` is
   called for the first time.
 - Bump up default size limit for `Accept-Language` and `User-Agent` to 2048